pyproject.toml
[tool.poetry]
name = "rouge"
version = "0.1.0"
description = ""
authors = ["Shineaki <halmosadam92@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
pygame-ce = "^2.5.2"
numpy = "^2.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
